Cleaned VidHD mode uniforms

+++ VidHdWindowBeam.h
@@ -138,7 +138,7 @@ private:
 	glm::ivec2 modeSize = glm::ivec2(0, 0);	// Row and Column count
 	int fontTex = _TEXUNIT_IMAGE_FONT_ROM_DEFAULT - GL_TEXTURE0;
 	glm::uvec2 glyphSize = glm::uvec2(14,16);
-	glm::vec2 fontScale = glm::vec2(2.0,2.0);	// Font size should be 16x16
+	glm::uvec2 fontScale = glm::vec2(2,2);		// Font size should be 16x16
 	SDL_FRect quad = { -1.f, 1.f, 2.f, -2.f };	// x, y, width, height
 
 	void UpdateVertexArray();
